requirements.txt
Appium-Python-Client==5.1.1
async-generator==1.10
attrs==25.3.0
certifi==2025.4.26
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==36.0.2
exceptiongroup==1.3.0
h11==0.13.0
idna==3.3
iniconfig==2.1.0
numpy==1.22.1
outcome==1.3.0.post0
packaging==25.0
pandas==1.3.5
pluggy==1.6.0
pycparser==2.21
Pygments==2.19.1
pyOpenSSL==22.0.0
PySocks==1.7.1
pytest==8.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0.2
requests==2.27.1
selenium==4.33.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==3.5.4
websocket-client==1.8.0
wsproto==1.1.0
